mp-gemini/requirements.txt
streamlit
python-dotenv

langchain
langchain-openai
langchain-core
langchain-community

chromadb
tiktoken

pypdf

mcp
# SDK oficial do Databricks
databricks-sdk
# Interface gráfica

# Utilitários assíncronos (necessários para o cliente MCP no Streamlit)
asyncio